+++ sdk/resourcemanager/keyvault/armkeyvault/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 2.0.0 (2024-07-18)
+### Breaking Changes
+
+- Function `*VaultsClient.BeginCreateOrUpdate` has been removed
+- Function `*VaultsClient.Delete` has been removed
+- Function `*VaultsClient.Get` has been removed
+- Function `*VaultsClient.Update` has been removed
+- Struct `VaultCreateOrUpdateParameters` has been removed
+- Struct `VaultPatchParameters` has been removed
+- Struct `VaultPatchProperties` has been removed
+
+
 ## 1.4.0 (2023-11-24)
 ### Features Added